fix build 3

+++ tasks/mpi/volochaev_s_count_characters_27/src/ops_mpi.cpp
@@ -77,12 +77,12 @@ bool volochaev_s_count_characters_27_mpi::Lab1_27_mpi::pre_processing() {
   if (world.rank() == 0) {
     // Init vectors
     input_ = std::vector<std::pair<char, char>>(world.size() * delta);
-    for (unsigned i = 0; i < std::min(tmp1.size(), tmp2.size()); i++) {
+    for (size_t i = 0; i < std::min(tmp1.size(), tmp2.size()); i++) {
       input_[i].first = tmp1[i];
       input_[i].second = tmp2[i];
     }
 
-    for (size_t proc = 1; proc < world.size(); proc++) {
+    for (int proc = 1; proc < world.size(); proc++) {
       world.send(proc, 0, input_.data() + proc * delta, delta);
     }
   }
